Rename extension method classes
