+++ Tests/Testably.Architecture.Testing.Tests/ExtensionsForMethodInfoTests.cs
@@ -3,9 +3,9 @@ using System;
 using System.Reflection;
 using Xunit;
 
-namespace Testably.Architecture.Testing.Tests.Extensions;
+namespace Testably.Architecture.Testing.Tests;
 
-public sealed class MethodInfoTests
+public sealed class ExtensionsForMethodInfoTests
 {
 	[Fact]
 	public void HasAttribute_WithAttribute_ShouldReturnTrue()
