+++ Tests/Testably.Architecture.Testing.Tests/Internal/TestResultTests.cs
@@ -4,7 +4,7 @@ using System.Reflection;
 using Testably.Architecture.Testing.TestErrors;
 using Xunit;
 
-namespace Testably.Architecture.Testing.Tests;
+namespace Testably.Architecture.Testing.Tests.Internal;
 
 public sealed class TestResultTests
 {
